refactor(StackedChart): rename StackedChart to CumulativeChart
