+++ src/components/CumulativeChart.vue
